Improve duplicate field error message
Provide a better error message if identical fields with different case
are provided as source and target. For example, 'LastEditor' vs.
'LASTEDITOR'. This is not supported by databases. This should only
happen when the field already exists, like 'Linetype' from AutoCAD and
the field mapping includes 'LINETYPE' as a new field name. Other cases
are already handled.